+++ examples/environment.rs
@@ -1,7 +1,3 @@
-#[macro_use]
-extern crate log;
-extern crate simple_logger;
-
 use std::error::Error;
 
 fn main() -> Result<(), Box<dyn Error>> {
@@ -13,7 +9,7 @@ fn main() -> Result<(), Box<dyn Error>> {
 
     spawn("export");
 
-    sudo::with_env(&["EXAMPLE_", "CARGO"])?;
+    sudo2::with_env(&["EXAMPLE_", "CARGO"])?;
 
     uid_euid("②");
 
@@ -25,12 +21,12 @@ fn main() -> Result<(), Box<dyn Error>> {
 fn uid_euid(nth: &str) {
     let euid = unsafe { libc::geteuid() };
     let uid = unsafe { libc::getuid() };
-    info!("{} uid: {}; euid: {};", nth, uid, euid);
+    tracing::info!("{} uid: {}; euid: {};", nth, uid, euid);
 }
 
 fn spawn(cmd: &str) {
     let mut child = std::process::Command::new("/usr/bin/env")
-        .args(&["bash", "-c", cmd])
+        .args(["bash", "-c", cmd])
         .spawn()
         .expect("unable to start child");