+++ src/lib.rs
@@ -1,20 +1,20 @@
-//! [![crates.io](https://img.shields.io/crates/v/sudo?logo=rust)](https://crates.io/crates/karen/)
+//! [![crates.io](https://img.shields.io/crates/v/sudo?logo=rust)](https://crates.io/crates/sudo/)
 //! [![docs.rs](https://docs.rs/sudo/badge.svg)](https://docs.rs/sudo)
 //!
-//! Detect if you are running as root, restart self with `sudo` if needed or setup uid zero when running with the SUID flag set.
+//! Detect if you are running as root, restart self with `sudo` if needed or
+//! setup uid zero when running with the SUID flag set.
 //!
 //! ## Requirements
 //!
-//! * The `sudo` program is required to be installed and setup correctly on the target system.
+//! * The `sudo` program is required to be installed and setup correctly on the
+//!   target system.
 //! * Linux or Mac OS X tested
-//!     * It should work on *BSD. However, you can also create an Escalate builder with `doas` as the wrapper should you prefer that.
+//!     * It should work on *BSD. However, you can also create an Escalate
+//!       builder with `doas` as the wrapper should you prefer that.
 #![allow(clippy::bool_comparison)]
 
 use std::error::Error;
 use std::process::Command;
-
-#[macro_use]
-extern crate log;
 
 /// Cross platform representation of the state the current program running
 #[derive(Debug, PartialEq)]
@@ -23,15 +23,17 @@ pub enum RunningAs {
     Root,
     /// Running as a normal user
     User,
-    /// Started from SUID, a call to `sudo::escalate_if_needed` or `karen::with_env` is required to claim the root privileges at runtime.
+    /// Started from SUID, a call to `sudo2::escalate_if_needed` or
+    /// `sudo2::with_env` is required to claim the root privileges at runtime.
     /// This does not restart the process.
     Suid,
 }
 use RunningAs::*;
 
 #[cfg(unix)]
-/// Check getuid() and geteuid() to learn about the configuration this program is running under
-pub fn check() -> RunningAs {
+/// Check getuid() and geteuid() to learn about the configuration this program
+/// is running under
+fn check() -> RunningAs {
     let uid = unsafe { libc::getuid() };
     let euid = unsafe { libc::geteuid() };
 
@@ -41,6 +43,17 @@ pub fn check() -> RunningAs {
         (_, _) => User,
     }
     //if uid == 0 { Root } else { User }
+}
+
+#[cfg(unix)]
+/// Returns `true` if binary is already running as root.
+pub fn running_as_root() -> bool {
+    check() == RunningAs::Root
+}
+
+/// Returns `true` if binary is already running as suid.
+pub fn running_as_suid() -> bool {
+    check() == RunningAs::Suid
 }
 
 pub struct Escalate {
@@ -56,35 +69,49 @@ impl Default for Escalate {
 }
 
 impl Escalate {
-    pub fn builder() -> Self {
+    fn builder() -> Self {
         Default::default()
     }
 
-    pub fn wrapper(&mut self, wrapper: &str) -> &mut Self {
+    fn wrapper(&mut self, wrapper: &str) -> &mut Self {
         self.wrapper = wrapper.to_string();
         self
     }
 
-    ///  Escalate privileges while maintaining RUST_BACKTRACE and selected environment variables (or none).
+    /// Escalate privileges while maintaining RUST_BACKTRACE and selected
+    /// environment variables (or none).
     ///
     /// Activates SUID privileges when available.
-    pub fn with_env(&self, prefixes: &[&str]) -> Result<RunningAs, Box<dyn Error>> {
+    fn with_env(&self, prefixes: &[&str]) -> Result<RunningAs, Box<dyn Error>> {
+        self.collect_envs(prefixes, false)
+    }
+
+    /// Escalate privileges while maintaining RUST_BACKTRACE and selected
+    /// environment variables (or none) as wildcard. Use can use `*` to select
+    /// all environment variables (mimics `sudo -E`)
+    ///
+    /// Activates SUID privileges when available.
+    fn with_env_wildcards(&self, wildcards: &[&str]) -> Result<RunningAs, Box<dyn Error>> {
+        self.collect_envs(wildcards, true)
+    }
+
+    fn collect_envs(&self, patterns: &[&str], is_glob: bool) -> Result<RunningAs, Box<dyn Error>> {
         let current = check();
-        trace!("Running as {:?}", current);
+        tracing::trace!("Running as {:?}", current);
         match current {
             Root => {
-                trace!("already running as Root");
+                tracing::trace!("already running as Root");
                 return Ok(current);
             }
             Suid => {
-                trace!("setuid(0)");
+                tracing::trace!("setuid(0)");
                 unsafe {
                     libc::setuid(0);
                 }
                 return Ok(current);
             }
             User => {
-                debug!("Escalating privileges");
+                tracing::debug!("Escalating privileges");
             }
         }
 
@@ -104,7 +131,7 @@ impl Escalate {
                 "1" | "true" => Some("1"),
                 "full" => Some("full"),
                 invalid => {
-                    warn!(
+                    tracing::warn!(
                         "RUST_BACKTRACE has invalid value {:?} -> defaulting to \"full\"",
                         invalid
                     );
@@ -112,21 +139,31 @@ impl Escalate {
                 }
             };
             if let Some(value) = value {
-                trace!("relaying RUST_BACKTRACE={}", value);
+                tracing::trace!("relaying RUST_BACKTRACE={}", value);
                 // command.arg(format!("RUST_BACKTRACE={}", value));
                 command.env("RUST_BACKTRACE", value);
             }
         }
 
-        if !prefixes.is_empty() {
+        if !patterns.is_empty() {
             // Only add env for pkexec if we're passing any additional env vars
             if self.wrapper == "pkexec" {
-                trace!("Prefixing `env` to pkexec command to pass additional environment variables! This may break pkexec system policies.");
+                tracing::trace!(
+                    "Prefixing `env` to pkexec command to pass additional environment variables! \
+                     This may break pkexec system policies."
+                );
                 command.arg("env");
             }
+
             for (name, value) in std::env::vars().filter(|(name, _)| name != "RUST_BACKTRACE") {
-                if prefixes.iter().any(|prefix| name.starts_with(prefix)) {
-                    trace!("propagating {}={}", name, value);
+                if patterns.iter().any(|pattern| {
+                    if is_glob {
+                        wildmatch::WildMatch::new(pattern).matches(&name)
+                    } else {
+                        name.starts_with(pattern)
+                    }
+                }) {
+                    tracing::trace!("propagating {}={}", name, value);
                     if self.wrapper == "pkexec" {
                         command.arg(format!("{}={}", name, value));
                     }
@@ -146,7 +183,8 @@ impl Escalate {
         }
     }
 
-    /// Restart your program with root privileges if the user is not privileged enough.
+    /// Restart your program with root privileges if the user is not privileged
+    /// enough.
     ///
     /// Activates SUID privileges when available
     pub fn escalate_if_needed(&self) -> Result<RunningAs, Box<dyn Error>> {
@@ -168,8 +206,8 @@ pub fn builder() -> Escalate {
 /// ```
 /// # use std::error::Error;
 /// # fn main() -> Result<(), Box<dyn Error>> {
-/// #   if sudo::check() == karen::RunningAs::Root {
-/// sudo::escalate_if_needed()?;
+/// #   if sudo2::running_as_root() {
+/// sudo2::escalate_if_needed()?;
 /// // the following gets only executed in privileged mode
 /// #   } else {
 /// #     eprintln!("not actually testing");
@@ -188,8 +226,8 @@ pub fn escalate_if_needed() -> Result<RunningAs, Box<dyn Error>> {
 /// ```
 /// # use std::error::Error;
 /// # fn main() -> Result<(), Box<dyn Error>> {
-/// #   if sudo::check() == karen::RunningAs::Root {
-/// sudo::pkexec()?;
+/// #   if sudo2::running_as_root() {
+/// sudo2::pkexec()?;
 /// // the following gets only executed in privileged mode
 /// #   } else {
 /// #     eprintln!("not actually testing");
@@ -208,8 +246,8 @@ pub fn pkexec() -> Result<RunningAs, Box<dyn Error>> {
 /// ```
 /// # use std::error::Error;
 /// # fn main() -> Result<(), Box<dyn Error>> {
-/// #   if sudo::check() == karen::RunningAs::Root {
-/// sudo::doas()?;
+/// #   if sudo2::running_as_root() {
+/// sudo2::doas()?;
 /// // the following gets only executed in privileged mode
 /// #   } else {
 /// #     eprintln!("not actually testing");
@@ -223,15 +261,16 @@ pub fn doas() -> Result<RunningAs, Box<dyn Error>> {
 }
 
 #[cfg(unix)]
-/// Escalate privileges while maintaining RUST_BACKTRACE and selected environment variables (or none).
+/// Escalate privileges while maintaining RUST_BACKTRACE and selected
+/// environment variables (or none).
 ///
 /// Activates SUID privileges when available.
 ///
 /// ```
 /// # use std::error::Error;
 /// # fn main() -> Result<(), Box<dyn Error>> {
-/// #   if sudo::check() == karen::RunningAs::Root {
-/// sudo::with_env(&["CARGO_", "MY_APP_"])?;
+/// #   if sudo2::running_as_root() {
+/// sudo2::with_env(&["CARGO_", "MY_APP_"])?;
 /// // the following gets only executed in privileged mode
 /// #   } else {
 /// #     eprintln!("not actually testing");
@@ -243,13 +282,66 @@ pub fn with_env(prefixes: &[&str]) -> Result<RunningAs, Box<dyn Error>> {
     Escalate::default().with_env(prefixes)
 }
 
+#[cfg(unix)]
+/// Escalate privileges while maintaining RUST_BACKTRACE and selected
+/// environment variables that matches given wildcard (or none).
+///
+/// To select all env variables, use `*`. Note that it may be insecure. Use it
+/// with care.
+///
+/// Activates SUID privileges when available.
+///
+/// ```
+/// # use std::error::Error;
+/// # fn main() -> Result<(), Box<dyn Error>> {
+/// #   if sudo2::running_as_root() {
+/// sudo2::with_env_wildcards(&["CARGO_*", "MY_APP_*"])?;
+/// // the following gets only executed in privileged mode
+/// #   } else {
+/// #     eprintln!("not actually testing");
+/// #   }
+/// #   Ok(())
+/// # }
+/// ```
+pub fn with_env_wildcards(wildcards: &[&str]) -> Result<RunningAs, Box<dyn Error>> {
+    Escalate::default().with_env_wildcards(wildcards)
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
+    use tracing_test::traced_test;
 
     #[test]
+    #[traced_test]
     fn it_works() {
         let c = check();
-        assert!(true, "{:?}", c);
+        println!("{:?}", c);
+    }
+
+    #[test]
+    #[traced_test]
+    fn sudo_with_env() {
+        std::env::set_var("CARGO_FOO", "1");
+        std::env::set_var("CARGO_BAR_BAZ", "1");
+        with_env(&["CARGO_"]).unwrap();
+
+        let mut vars = std::env::vars();
+        assert!(vars.any(|(k, _v)| k == "CARGO_FOO"));
+        assert!(vars.any(|(k, _v)| k == "CARGO_BAR_BAZ"));
+        assert!(!vars.any(|(k, _v)| k == "CARGO_FOO_BAR_BAZ"));
+    }
+
+    #[test]
+    #[traced_test]
+    fn sudo_with_env_wildcard() {
+        std::env::set_var("CARGO_FOO", "1");
+        std::env::set_var("CARGO_BAR_BAZ", "1");
+        with_env_wildcards(&["CARGO_*"]).unwrap();
+
+        let mut vars = std::env::vars();
+        assert!(vars.any(|(k, _v)| k == "CARGO_FOO"));
+        assert!(vars.any(|(k, _v)| k == "CARGO_BAR_BAZ"));
+        assert!(!vars.any(|(k, _v)| k == "CARGO_FOO_BAR_BAZ"));
     }
 }